Library/LibraryTests/geminiTests/many/BorrowTest.cs
using Library.files.resources;


/* 
 * Usystematyzowanie namespace
 * Poniżej został zakomentowany wygenerowany kod
 * Poprawki w kodzie:
 * Dodanie dodatkowej przestrzeni nazw
 * Usunięcie Assertów, w których czat próbuje uzyskać dostęp do prywatnych pól
 */

namespace Library.Tests.gemini.many;
public class BorrowTests
{

}

/*
using Library.files.resources;

[TestClass]
public class BorrowTests
{
    [TestMethod]
    public void AddBook_ShouldIncrementBookID()
    {
        // Arrange
        var borrow = new Borrow();

        // Act
        borrow.AddBook("Title 1", "Author 1", 2023);
        borrow.AddBook("Title 2", "Author 2", 2024);

        // Assert
        Assert.AreEqual(2, borrow.Books.Count);
        Assert.AreEqual(2, borrow.Books[1].GetID());
    }

    // ... other test methods ...

    [TestMethod]
    public void AddBook_WithEmptyTitle_ShouldThrowException()
    {
        // Arrange
        var borrow = new Borrow();

        // Act & Assert
        Assert.ThrowsException<ArgumentException>(() => borrow.AddBook("", "Author", 2023));
    }

    [TestMethod]
    public void BorrowBook_WithInvalidUserID_ShouldReturnFalse()
    {
        // ...
    }
}
*/